backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydicom==2.4.3
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
groq
python-dotenv==1.0.0
requests==2.31.0
roboflow==1.1.25
httpx==0.24.1
gdcm
pylibjpeg
pylibjpeg-libjpeg
aiofiles==23.2.1
pytest